Added pre-existing .py files from another repo I was working in.
Using tkinter for GUI, there are other options but this seemed the most appealing to use. May change in the future.

Would like to state that I have not done this before and I am using a template online that I've found. Will be dissecting what I've learnt and apply it by myself.