Fixed an overlooked problem
Thinking that I had to create an entire new function to handle new operations on pre-existing results, the solution was much simpler. Global variable operations is the variable that gets constantly used to for whatever state the calculator is pretty much in, by assigning the result variable I had already made, I can then add further operations if I wanted to.

With that, I have made so that all operations is now part of the Page since it's a member variable, instead of calling it as "Operations.window_result.xyz".